requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
oauthlib==3.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
yapf==0.30.0
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
scipy>=0.15.1
pandas>=0.19
regex==2020.11.13
requests-oauthlib==1.3.0
tweepy==3.10.0
Unidecode==1.1.2
wordcloud==1.8.1
textblob==0.15.3
olefile==0.46
